+++ run-time-error.d.ts
@@ -14,7 +14,7 @@
  * }
  * ```
  */
-export default class RuntimeError extends Error
+export class RuntimeError extends Error
 {
 	/**
 	 * Creates a runtime error.